# backend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "TeamPrompts.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5258/"
    }
  ],
  "paths": {
...
    "/api/scripts": {
      "get": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScriptListItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScriptListItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScriptListItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Scripts"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "workspaceId": {
                        "type": "string",
                        "format": "uuid"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "ContentType": {
                        "type": "string"
                      },
                      "ContentDisposition": {
                        "type": "string"
                      },
                      "Headers": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "Length": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "format": "int64"
                      },
                      "Name": {
                        "type": "string"
                      },
                      "FileName": {
                        "type": "string"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/scripts/{id}": {
      "get": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScriptRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScriptRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScriptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/scripts/{id}/sessions": {
      "get": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionWithResultsDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionWithResultsDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionWithResultsDto"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/scripts/{id}/tray": {
      "get": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrayItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrayItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrayItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/scripts/{id}/canvas": {
      "get": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CanvasNodeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CanvasNodeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CanvasNodeDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCanvasRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCanvasRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCanvasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Scripts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanvasNodeDto": {
        "required": [
          "nodeKey",
          "x",
          "y"
        ],
        "type": "object",
        "properties": {
          "nodeKey": {
            "type": "string"
          },
          "x": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "format": "double"
          },
          "y": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "format": "double"
          }
        }
      },
...
      "FileType": {
        "enum": [
          "Pdf",
          "Txt"
        ]
      },
...
      "GenerationResultDto": {
        "required": [
          "id",
          "sessionId",
          "index",
          "content",
          "kind",
          "createdAt",
          "isFavorite",
          "favoriteCount",
          "copyCount",
          "isHighlighted",
          "highlightedBy",
          "highlightedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "index": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "content": {
            "type": "string"
          },
          "kind": {
            "oneOf": [
              {
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ResultKind"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "favoriteCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "copyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "isHighlighted": {
            "type": "boolean"
          },
          "highlightedBy": {
            "oneOf": [
              {
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/UserRef"
              }
            ]
          },
          "highlightedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "ResultKind": {
        "enum": [
          "Title",
          "Description",
          "Hook",
          "Tags",
          "Thumbnail",
          null
        ]
      },
      "SaveCanvasRequest": {
        "required": [
          "nodes"
        ],
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CanvasNodeDto"
            }
          }
        }
      },
      "ScriptDto": {
        "required": [
          "id",
          "name",
          "originalFileName",
          "fileType",
          "extractedText",
          "createdAt",
          "updatedAt",
          "createdBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "fileType": {
            "$ref": "#/components/schemas/FileType"
          },
          "extractedText": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserRef"
          }
        }
      },
      "ScriptListItemDto": {
        "required": [
          "id",
          "name",
          "originalFileName",
          "fileType",
          "createdAt",
          "updatedAt",
          "createdBy",
          "sessionCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "fileType": {
            "$ref": "#/components/schemas/FileType"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserRef"
          },
          "sessionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          }
        }
      },
      "SessionDto": {
        "required": [
          "id",
          "runId",
          "scriptId",
          "promptId",
          "promptVersionId",
          "promptName",
          "model",
          "status",
          "error",
          "createdBy",
          "createdAt",
          "completedAt",
          "promptVersionNumber",
          "isMainVersion",
          "promptVersionNote"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "runId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "scriptId": {
            "type": "string",
            "format": "uuid"
          },
          "promptId": {
            "type": "string",
            "format": "uuid"
          },
          "promptVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "promptName": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserRef"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "promptVersionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "isMainVersion": {
            "type": "boolean"
          },
          "promptVersionNote": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SessionStatus": {
        "enum": [
          "Queued",
          "Streaming",
          "Completed",
          "Failed"
        ]
      },
      "SessionWithResultsDto": {
        "required": [
          "session",
          "results"
        ],
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionDto"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationResultDto"
            }
          }
        }
      },
...
      "TrayItemDto": {
        "required": [
          "resultId",
          "sessionId",
          "content",
          "kind",
          "promptName",
          "model",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "resultId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "kind": {
            "oneOf": [
              {
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ResultKind"
              }
            ]
          },
          "promptName": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateScriptRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "UserRef": {
        "required": [
          "id",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}